utils: fix int64 hex conversion e70f759b added a guard which works fine if 'hex' is equal or less than 32bit. This fix is to support when 'hex' is longer than 32bit and less than int64_t.

+++ src/flb_utils.c
@@ -455,7 +455,7 @@ int64_t flb_utils_hex2int(char *hex, int len)
 
     while ((c = *hex++) && i < len) {
         /* Ensure no overflow */
-        if (res >= (int)((INT64_MAX/0x10) - 0xff)) {
+        if (res >= (int64_t)((INT64_MAX/0x10) - 0xff)) {
             return -1;
         }
